+++ r2dbc/mysql/src/main/java/com/google/cloud/sql/core/GcpConnectionFactoryProviderMysql.java
@@ -47,8 +47,9 @@ public class GcpConnectionFactoryProviderMysql extends GcpConnectionFactoryProvi
   }
 
   @Override
-  ConnectionFactory tcpConnectonFactory(
+  ConnectionFactory tcpConnectionFactory(
       Builder optionBuilder,
+      String ipTypes,
       Function<SslContextBuilder, SslContextBuilder> customizer,
       String csqlHostName) {
     optionBuilder
@@ -58,6 +59,7 @@ public class GcpConnectionFactoryProviderMysql extends GcpConnectionFactoryProvi
         .option(MySqlConnectionFactoryProvider.TCP_KEEP_ALIVE, true);
     return new CloudSqlConnectionFactory(
         (ConnectionFactoryOptions options) -> new MySqlConnectionFactoryProvider().create(options),
+        ipTypes,
         optionBuilder,
         csqlHostName);
   }