+++ r2dbc/postgres/src/main/java/com/google/cloud/sql/core/GcpConnectionFactoryProviderPostgres.java
@@ -52,8 +52,9 @@ public class GcpConnectionFactoryProviderPostgres extends GcpConnectionFactoryPr
   }
 
   @Override
-  ConnectionFactory tcpConnectonFactory(
+  ConnectionFactory tcpConnectionFactory(
       Builder optionBuilder,
+      String ipTypes,
       Function<SslContextBuilder, SslContextBuilder> customizer,
       String csqlHostName) {
     optionBuilder
@@ -64,6 +65,7 @@ public class GcpConnectionFactoryProviderPostgres extends GcpConnectionFactoryPr
     return new CloudSqlConnectionFactory(
         (ConnectionFactoryOptions options) ->
             new PostgresqlConnectionFactoryProvider().create(options),
+        ipTypes,
         optionBuilder,
         csqlHostName);
   }