+++ r2dbc/sqlserver/src/main/java/com/google/cloud/sql/core/GcpConnectionFactoryProviderMssql.java
@@ -46,8 +46,9 @@ public class GcpConnectionFactoryProviderMssql extends GcpConnectionFactoryProvi
   }
 
   @Override
-  ConnectionFactory tcpConnectonFactory(
+  ConnectionFactory tcpConnectionFactory(
       Builder optionBuilder,
+      String ipTypes,
       Function<SslContextBuilder, SslContextBuilder> customizer,
       String csqlHostName) {
     optionBuilder
@@ -57,6 +58,7 @@ public class GcpConnectionFactoryProviderMssql extends GcpConnectionFactoryProvi
 
     return new CloudSqlConnectionFactory(
         (ConnectionFactoryOptions options) -> new MssqlConnectionFactoryProvider().create(options),
+        ipTypes,
         optionBuilder,
         csqlHostName);
   }